+++ src/Instance/InstanceEdit/InstanceEdit.js
@@ -31,6 +31,8 @@ import useLoadSubInstances from '../../hooks/useLoadSubInstances';
 import useCallout from '../../hooks/useCallout';
 import { useInstanceMutation } from '../../hooks';
 
+import css from './InstanceEdit.css';
+
 const InstanceEdit = ({
   instanceId,
   referenceData,
@@ -74,7 +76,20 @@ const InstanceEdit = ({
   const onError = async error => {
     const response = await error.response;
     const parsedError = await parseHttpError(response);
-    setHttpError(parsedError);
+    const err = {
+      message: parsedError?.errors[0]?.message,
+      status: response.status,
+    };
+
+    setHttpError(err);
+  };
+
+  const getErrorModalContent = () => {
+    return (
+      <div className={css.errorModalContent}>
+        {httpError?.status ? `${httpError.status}: ${httpError.message}` : httpError.message}
+      </div>
+    );
   };
 
   const isMemberTenant = checkIfUserInMemberTenant(stripes);
@@ -107,7 +122,7 @@ const InstanceEdit = ({
         <ErrorModal
           open
           label={<FormattedMessage id="ui-inventory.instance.saveError" />}
-          content={httpError?.status ? `${httpError.status}: ${httpError.message}` : httpError.message}
+          content={getErrorModalContent()}
           onClose={() => setHttpError()}
         />
       }